Drop unneeded testcryptoutils.h
